rapidocr/requirements.txt
flask>=3.0.0
rapidocr>=3.0.0
onnxruntime>=1.20.0
Pillow>=10.2.0
numpy>=1.24.0
